requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pyodbc
pydantic-settings
